requirements.txt
moviepy
requests
PyMuPDF
scholarly
pytesseract
pdf2image
nltk
scikit-learn
aiohttp
transformers
gtts
Pillow
tesseract
ffmpeg